requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.7.0
Pillow>=9.4.0
PyYAML>=6.0
tensorboard>=2.12.0
gdown>=4.7.0
timm>=0.9.0
